Fix a potential integer overflow
multiplication of two 32-bit integers will yield a 32-bit integer, and
that could overflow, even though it's gonna be stored in a 64-bit
integer. The overflow happens before widening.

+++ src/core/t-struct.c
@@ -585,7 +585,7 @@ static REBOOL assign_scalar(REBSTU *stu,
 			STATIC_ASSERT(sizeof(field->size) <= 4);
 			STATIC_ASSERT(sizeof(field->dimension) <= 4);
 
-			u64 step = field->size * field->dimension;
+			u64 step = (u64)field->size * (u64)field->dimension;
 			if (step > VAL_STRUCT_LIMIT) {
 				Trap1(RE_SIZE_LIMIT, out);
 			}
